A little bit of tidying, linting and compat
